+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - PHP can be detected from the Symfony profiler.
+- Apache httpd can be detected from its 404 page even if OpenSSL is listed.
 
 ## [1.2.0]